updated: moved component registration methods from CommandRegistry class to ComponentRegistry class
