+++ api/src/main/java/com/github/azzerial/slash/internal/CommandRegistry.java
@@ -19,13 +19,12 @@ package com.github.azzerial.slash.internal;
 import com.github.azzerial.slash.SlashCommand;
 import com.github.azzerial.slash.annotations.Slash;
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
-import net.dv8tion.jda.api.events.interaction.SelectionMenuEvent;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
 
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.util.*;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 
 public final class CommandRegistry {
 
@@ -61,8 +60,7 @@ public final class CommandRegistry {
     public SlashCommand registerCommand(Object obj) {
         final SlashCommand command = compileCommand(obj);
 
-        registerButtons(obj);
-        registerSelectionMenus(obj);
+        ComponentRegistry.getInstance().registerComponent(obj);
         registry.put(command.getTag(), command);
         return command;
     }
@@ -87,45 +85,5 @@ public final class CommandRegistry {
         final CommandData data = annotationCompiler.compileCommand(command);
         final Map<String, Method> handlers = annotationCompiler.compileHandlers(cls, data);
         return new SlashCommand(jda, tag.value(), data, obj, handlers);
-    }
-
-    private void registerButtons(Object obj) {
-        final Class<?> cls = obj.getClass();
-
-        Arrays.stream(cls.getDeclaredMethods())
-            .filter(method ->
-                (method.getModifiers() & (Modifier.PROTECTED | Modifier.PRIVATE)) == 0
-                    && method.isAnnotationPresent(Slash.Button.class)
-                    && method.getParameterCount() == 1
-                    && method.getParameterTypes()[0] == ButtonClickEvent.class
-            )
-            .sorted(Comparator.comparing(Method::getName))
-            .forEach(method -> {
-                final String tag = method.getAnnotation(Slash.Button.class).value();
-
-                if (!tag.isEmpty()) {
-                    ComponentRegistry.getInstance().registerComponent(tag, new ComponentCallback(obj, method));
-                }
-            });
-    }
-
-    private void registerSelectionMenus(Object obj) {
-        final Class<?> cls = obj.getClass();
-
-        Arrays.stream(cls.getDeclaredMethods())
-            .filter(method ->
-                (method.getModifiers() & (Modifier.PROTECTED | Modifier.PRIVATE)) == 0
-                    && method.isAnnotationPresent(Slash.SelectionMenu.class)
-                    && method.getParameterCount() == 1
-                    && method.getParameterTypes()[0] == SelectionMenuEvent.class
-            )
-            .sorted(Comparator.comparing(Method::getName))
-            .forEach(method -> {
-                final String tag = method.getAnnotation(Slash.SelectionMenu.class).value();
-
-                if (!tag.isEmpty()) {
-                    ComponentRegistry.getInstance().registerComponent(tag, new ComponentCallback(obj, method));
-                }
-            });
     }
 }
